Client Side Implementation
It is built using NUXT.js, an extension to Vue.js framework.Each UI elements is a separate custom component. Different pages uses these components to display information. Vuetify was a framework used to theme and enhance the design of the website, and vue-charts (a Vue-wrapper for charts.js) was used to render the necessary charts.